refactor(app.js): removed app as a serverless function endpoint
Renamed app/index.js to app/app.js so that it doesn't appear in the least of available serverless
functions in Netlify console.
