+++ functions/api.js
@@ -3,7 +3,7 @@
  * @see https://www.fastify.io/docs/latest/Serverless/#lambdajs
  */
 const awsLambdaFastify = require('aws-lambda-fastify');
-const init = require('./app');
+const init = require('./app/app');
 
 const proxy = awsLambdaFastify(init());
 exports.handler = proxy;
